requirements.txt
--index-url https://pypi.fury.io/zerthmonk/
--extra-index-url https://pypi.org/simple/

attrs>=19.1.0
Django>=4.0
django-cors-headers>=3.1.0
django-filter>=2.2.0
djangorestframework>=3.10.3
djangorestframework-jwt>=1.11.0
django_extensions
django-polymorphic>=3.1.0
gunicorn
Pillow
PyJWT>=1.7.1
PyYAML>=5.1.2
pytz>=2019.2
mypy==0.730
psycopg2-binary
pytest-django>=3.5.1
pytest-mock>=1.11
deepdiff>=4.0.0
flake8
skabenproto
kombu
isort
vulture
black
requests

